+++ engine/renderer/Mesh.h
@@ -7,7 +7,7 @@ class Mesh
 {
 public:
     void Init(Device& device, CommandManager& cmdManager,const std::vector<Vertex>& vertices,
-        const std::vector<uint16_t>& indices);
+        const std::vector<uint32_t>& indices);
 
     void Shutdown();
